fix: remove stale fixtures/integration/ and rename golden to expected
Remove the duplicate fixtures/integration/ directory (tests read from
fynd-core/tests/fixtures/). Rename golden.rs to expected.rs and replace
all "golden" references with "expected" for consistent naming.
