--- fixtures/integration/README.md
@@ -1,92 +1,0 @@
-# Integration Test Fixtures
-
-This directory contains recorded market state and golden output baselines used
-by the fynd-core integration tests.
-
-## Files
-
-- `market_recording.json.zst` — Zstd-compressed JSON recording of Tycho
-  protocol stream updates. Captured from live Tycho mainnet using the
-  `record-market` tool.
-- `expected_outputs.json` — Expected quote outputs for a set of canonical
-  trading pairs. Generated by replaying the recording through the full Fynd
-  pipeline (the same `record-market` tool).
-
-## Recording Market State
-
-Use the `record-market` tool to capture a fresh recording and regenerate
-golden outputs in one step:
-
-```bash
-RUST_LOG=info cargo run -p record-market -- \
-  --tycho-url tycho-beta.propellerheads.xyz \
-  --tycho-api-key "$TYCHO_API_KEY" \
-  --rpc-url "$RPC_URL" \
-  --duration-secs 30 \
-  --output-dir fixtures/integration \
-  --min-tvl 10.0 \
-  --traded-n-days-ago 3
-```
-
-### CLI Options
-
-| Flag | Default | Description |
-|------|---------|-------------|
-| `--tycho-url` | (required) | Tycho WebSocket endpoint |
-| `--tycho-api-key` | (required) | Tycho API key |
-| `--rpc-url` | (optional) | Ethereum RPC URL for gas price capture |
-| `--duration-secs` | 600 | How long to record stream updates |
-| `--protocols` | auto-discover | Comma-separated protocol filter |
-| `--min-tvl` | 10.0 | Minimum TVL in ETH for pool filtering |
-| `--min-token-quality` | 100 | Minimum token quality score |
-| `--traded-n-days-ago` | 3 | Only include recently traded tokens |
-
-### Gas Price
-
-If `--rpc-url` is provided, the tool fetches the current gas price from the
-node and stores the raw wei value in the recording metadata. During replay,
-this recorded gas price is injected into the market data so the solver can
-compute gas cost deductions without needing a live RPC connection.
-
-If `--rpc-url` is omitted, a default of 10 gwei is used during replay.
-
-## Running Integration Tests
-
-```bash
-cargo test -p fynd-core --test integration
-```
-
-The tests replay the recorded market state through the full Fynd pipeline
-(TychoFeed → ComputationManager → WorkerPools → Router) and verify:
-
-- **Solution availability**: All pairs that succeeded in the golden baseline
-  also succeed during replay.
-- **Quality baseline**: Output amounts stay within 1% of the golden baseline.
-- **Quality invariants**: Successful quotes have positive output and gas
-  estimates.
-- **Derived data coverage**: Spot prices, pool depths, and token gas prices
-  cover expected thresholds.
-- **Timing**: Solve times stay within bounds derived from pool configuration.
-
-## Worker Pool Configuration
-
-Both the recording tool and integration tests read pool configuration from
-`worker_pools.toml` at the repository root. This ensures the golden baseline
-and test assertions use the same algorithm settings as production.
-
-## Limitations
-
-- **Ethereum only**: The recording tool and tests are currently hardcoded to
-  `Chain::Ethereum`. Multi-chain support would require parameterizing the
-  chain throughout the pipeline.
-- **VM-backed protocols filtered**: Protocol states that depend on EVM engine
-  state (e.g., UniswapV4) cannot be serialized and are dropped during
-  recording. Pools for these protocols are registered as components but have
-  no simulation state, so they cannot compute spot prices or be used for
-  routing.
-- **No live RPC during replay**: Gas prices must be recorded or default to
-  10 gwei. Token prices computed from replayed data may differ slightly from
-  live conditions.
-- **Recording size**: A 30-second recording produces ~700KB compressed.
-  Longer recordings capture more block updates but the initial snapshot
-  (first update) contains most of the state.
