+++ tools/record-market/src/expected.rs
